Move cookie store to a separate package
