+++ cookie/cookie.go
@@ -1,11 +1,12 @@
-package sessions
+package cookie
 
 import (
-	"github.com/gorilla/sessions"
+	"github.com/gin-contrib/sessions"
+	gsessions "github.com/gorilla/sessions"
 )
 
-type CookieStore interface {
-	Store
+type Store interface {
+	sessions.Store
 }
 
 // Keys are defined in pairs to allow key rotation, but the common case is to set a single
@@ -17,16 +18,16 @@ type CookieStore interface {
 //
 // It is recommended to use an authentication key with 32 or 64 bytes. The encryption key,
 // if set, must be either 16, 24, or 32 bytes to select AES-128, AES-192, or AES-256 modes.
-func NewCookieStore(keyPairs ...[]byte) CookieStore {
-	return &cookieStore{sessions.NewCookieStore(keyPairs...)}
+func NewStore(keyPairs ...[]byte) Store {
+	return &store{gsessions.NewCookieStore(keyPairs...)}
 }
 
-type cookieStore struct {
-	*sessions.CookieStore
+type store struct {
+	*gsessions.CookieStore
 }
 
-func (c *cookieStore) Options(options Options) {
-	c.CookieStore.Options = &sessions.Options{
+func (c *store) Options(options sessions.Options) {
+	c.CookieStore.Options = &gsessions.Options{
 		Path:     options.Path,
 		Domain:   options.Domain,
 		MaxAge:   options.MaxAge,
